requirements.txt
alembic==0.9.9
amqp==2.3.1
APScheduler==3.5.1
billiard==3.5.0.3
celery==4.1.1
click==6.7
cx-Oracle==6.3.1
Flask==1.0.2
Flask-APScheduler==1.8.0
Flask-Migrate==2.1.1
Flask-SQLAlchemy==2.3.2
itsdangerous==0.24
Jinja2==2.10
kombu==4.2.0
Mako==1.0.7
MarkupSafe==1.0
nmap==0.0.1
numpy==1.14.3
pandas==0.23.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
python-dateutil==2.7.3
python-dotenv==0.8.2
python-editor==1.0.3
python-nmap==0.6.1
pytz==2018.4
redis==2.10.6
six==1.11.0
SQLAlchemy==1.2.8
sqlparse==0.2.4
tzlocal==1.5.1
uWSGI==2.0.17
vine==1.1.4
Werkzeug==0.14.1
